refactor: 상수 파일을 관심사에 따라 여러 개의 상수 파일로 분리
