+++ src/components/addCardPage/SelectCardCompany/SelectCardCompany.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { CARD_COMPANY_KIND } from '../../../constants/constant';
+import { CARD_COMPANY_KIND } from '../../../constants/cardCompany';
 import { CardCompanyType } from '../../../types/types';
 import BankInfo from './SelectCardCompanyItem';
